Rename Tuesday(March 26).ipynb to Tuesday(March 26)LSTM.ipynb
this code is done with LSTM and a new argantinian dataset. the reason was to make sure that this task can be done. now that i have successfully done this with LSTM it might be also possible with transformer. also the dataset is very nice to use for later uses